+++ testing/conf.py
@@ -2,19 +2,7 @@
 # (c) 2022 Andreas Motl <andreas.motl@panodata.org>
 # License: GNU Affero General Public License, Version 3
 
-
-"""
-
 # Define example reading.
-from calypso_anemometer.model import CalypsoReading
+from boatface.model import DataValues
 
-test_reading = CalypsoReading(
-    wind_speed=5.69,
-    wind_direction=206,
-    battery_level=90,
-    temperature=33,
-    roll=30,
-    pitch=-60,
-    compass=235,
-)
-"""
+test_reading = DataValues(cog=42.42, dbt=84.84, sog=4.3, hdg=5.82, awa=42, aws=4.2, twa=170, tws=6.2)