+++ CHANGES.rst
@@ -9,6 +9,7 @@ in progress
 - Add data model classes and core display rendering routines
 - Add display subsystem, using one of PNG, pyglet, PySDL2, eips
 - Render all missing fields (AWA, TWA) and add corresponding assets
+- Packaging: Include assets into Python package
 
 
 2022-07-xx 0.0.0